make machine poller it's own timer

+++ server/Machines/MachineData.cs
@@ -2,8 +2,8 @@ namespace Eze.AdminConsole.Machines
 {
     public class MachineData
     {
-        public long idleCpuTime { get; internal set; }
-        public long cpuPercent { get; internal set; }
-        public long memoryMb { get; internal set; }
+        public float idleCpuTime { get; internal set; }
+        public float cpuPercent { get; internal set; }
+        public float memoryMb { get; internal set; }
     }
 }
